requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Talisman==1.1.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.8.2
argon2-cffi==23.1.0
psycopg[binary]>=3.1
pytest==8.3.2
gunicorn==22.0.0
PyJWT==2.9.0